Fix array items to also handle raw types
* Add handling parser.rb to handle constructs like { "type" => "string" }
* Fix ObjC parser to account for this (and move some functions into Utils modules)
* Update tests for parser and Obj properties
